+++ CHANGELOG.md
@@ -10,8 +10,20 @@
 - `withConfigurationCache` function
 - `withNoConfigurationCache` function
 - `withConfigurationCacheProblemsWarn` function
+- `withArgumentsFromTXT` function
+- `scan` and `addWithScan` functions
+- `info` and `addWithInfo` functions
+- `stacktrace` and `addWithStacktrace` functions
+- `buildCache` and `addWithBuildCache` functions
+- `noBuildCache`and `addWithNoBuildCache` functions
+- `configurationCache` and `andWithConfigurationCache` functions
+- `noConfigurationCache` and `andWithNoConfigurationCache` functions
+- `configurationCacheProblemsWarn` and `andWithConfigurationCacheProblemsWarn` functions
 
 ### Changed
+
+- all `with...` functions to `andWith...`
+- `gradleTestKitTest` function uses `withArgumentsFromTXT` flag instead of list of arguments
 
 ### Deprecated